Fix/improve the deck exporter/importer.
There were various problems with the plain text file importer that
caused it to fail to import cards if that card had certain characters in
its name, like a comma. Also, the incorrect side for a card got added.
Finally, there was no way to distinguish between cards that exist in
multiple sets with the same name. These all work now and required
modifying the export format slightly to make it easy to import to
another deck.